src/engine/requirements.txt
# Trading Forge - Python Engine Dependencies
# Core - required for all operations
numpy>=1.26.0
pandas>=2.2.0
polars>=1.0.0             # Primary data layer — 5-10x faster than Pandas for Parquet
pyarrow>=15.0.0           # Parquet support
duckdb>=1.0.0             # Query Parquet on S3 directly with SQL, zero-copy
scipy>=1.12.0             # Statistical functions, EVT, QMC, BCa intervals
vectorbt>=0.26.0          # Backtesting framework (Pandas boundary only)
pydantic>=2.6.0           # Config validation
click>=8.1.0              # CLI
optuna>=3.6.0             # Bayesian param search for robustness testing (TPE sampler)
requests>=2.31.0
boto3>=1.34.0             # AWS S3
databento>=0.40.0         # Historical market data

# Statistical / ML - required for backtesting pipeline
numba>=0.59.0             # JIT compilation for MC block bootstrap
arch>=7.0.0               # StationaryBootstrap, optimal_block_length, GARCH
hmmlearn>=0.3.2           # Hidden Markov Model regime detection
ruptures>=1.1.9           # PELT change-point detection
scikit-learn>=1.4.0       # LedoitWolf, MinCovDet, GraphicalLassoCV, KMeans

# Analytics
quantstats>=0.0.62        # QuantStats performance analytics
psutil>=5.9.0             # Hardware profiling

# Quantum (optional — graceful fallback to classical when unavailable)
# pip install qiskit qiskit-aer qiskit-algorithms dwave-samplers dwave-neal quimb pennylane
qiskit>=2.3.0
qiskit-aer>=0.15.0
qiskit-algorithms>=0.3.0
dwave-samplers>=1.7.0
dwave-neal>=0.6.0
quimb>=1.13.0
pennylane>=0.42.0

# Cloud QPU (optional)
qiskit-ibm-runtime>=0.46.0
amazon-braket-sdk>=1.90.0
pennylane-braket>=1.30.0

# DeepAR Regime Forecaster (local)
gluonts[torch]>=0.16.0
